+++ packages/app/src/routers/CRUDRouter/components/create-project/CreateProject.test.tsx
@@ -1,14 +1,16 @@
 /*---------------------------------------------------------------------------------------------
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
+ *
+ * This code is for demonstration purposes and should not be considered production ready.
  *--------------------------------------------------------------------------------------------*/
 import { toaster } from "@itwin/itwinui-react";
 import { act, fireEvent, render } from "@testing-library/react";
 import React from "react";
 
-import { UpdateProject } from "./UpdateProject";
+import { CreateProject } from "./CreateProject";
 
-describe("UpdateProject", () => {
+describe("CreateProject", () => {
   const mockedProject = { project: { id: "dd", displayName: "name" } };
   const fetchMock = jest.fn(() =>
     Promise.resolve({
@@ -26,44 +28,44 @@ describe("UpdateProject", () => {
     jest.clearAllMocks();
   });
 
-  it("should update a project", async () => {
+  it("should create a project", async () => {
     const successMock = jest.fn();
     toaster.positive = jest.fn();
 
     const { getByText, container } = render(
-      <UpdateProject
+      <CreateProject
         accessToken="dd"
-        projectId="de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b"
         onSuccess={successMock}
         apiOverrides={{ serverEnvironmentPrefix: "dev" }}
-        initialProject={{
-          displayName: "Initial name",
-          projectNumber: "Initial number",
-        }}
       />
     );
 
     const name = container.querySelector(
       "input[name=displayName]"
     ) as HTMLInputElement;
-    fireEvent.change(name, { target: { value: "Some other name" } });
+    fireEvent.change(name, { target: { value: "Some name" } });
 
-    const updateButton = getByText("Update");
-    await act(async () => updateButton.click());
+    const number = container.querySelector(
+      "input[name=projectNumber]"
+    ) as HTMLInputElement;
+    fireEvent.change(number, { target: { value: "Some number" } });
+
+    const createButton = getByText("Create");
+    await act(async () => createButton.click());
     expect(fetchMock).toHaveBeenCalledWith(
-      "https://dev-api.bentley.com/projects/de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b",
+      "https://dev-api.bentley.com/projects",
       {
-        method: "PATCH",
+        method: "POST",
         headers: { Authorization: "dd", Prefer: "return=representation" },
         body: JSON.stringify({
-          displayName: "Some other name",
-          projectNumber: "Initial number",
+          displayName: "Some name",
+          projectNumber: "Some number",
         }),
       }
     );
     expect(successMock).toHaveBeenCalledWith(mockedProject);
     expect(toaster.positive).toHaveBeenCalledWith(
-      "Project updated successfully.",
+      "Project created successfully.",
       {
         hasCloseButton: true,
       }
@@ -77,15 +79,10 @@ describe("UpdateProject", () => {
     toaster.negative = jest.fn();
 
     const { getByText, container } = render(
-      <UpdateProject
+      <CreateProject
         accessToken="dd"
-        projectId="de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b"
         onError={errorMock}
         apiOverrides={{ serverEnvironmentPrefix: "dev" }}
-        initialProject={{
-          displayName: "Initial name",
-          projectNumber: "Initial number",
-        }}
       />
     );
 
@@ -94,16 +91,21 @@ describe("UpdateProject", () => {
     ) as HTMLInputElement;
     fireEvent.change(name, { target: { value: "Some name" } });
 
-    const updateButton = getByText("Update");
-    await act(async () => updateButton.click());
+    const number = container.querySelector(
+      "input[name=projectNumber]"
+    ) as HTMLInputElement;
+    fireEvent.change(number, { target: { value: "Some number" } });
+
+    const createButton = getByText("Create");
+    await act(async () => createButton.click());
     expect(fetchMock).toHaveBeenCalledWith(
-      "https://dev-api.bentley.com/projects/de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b",
+      "https://dev-api.bentley.com/projects",
       {
-        method: "PATCH",
+        method: "POST",
         headers: { Authorization: "dd", Prefer: "return=representation" },
         body: JSON.stringify({
           displayName: "Some name",
-          projectNumber: "Initial number",
+          projectNumber: "Some number",
         }),
       }
     );
@@ -111,7 +113,7 @@ describe("UpdateProject", () => {
     expect(
       toaster.negative
     ).toHaveBeenCalledWith(
-      "Could not update a project. Please try again later.",
+      "Could not create a project. Please try again later.",
       { hasCloseButton: true }
     );
   });
@@ -123,15 +125,10 @@ describe("UpdateProject", () => {
     toaster.negative = jest.fn();
 
     const { getByText, container } = render(
-      <UpdateProject
+      <CreateProject
         accessToken="dd"
-        projectId="de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b"
         onError={errorMock}
         apiOverrides={{ serverEnvironmentPrefix: "dev" }}
-        initialProject={{
-          displayName: "Initial name",
-          projectNumber: "Initial number",
-        }}
       />
     );
 
@@ -140,16 +137,21 @@ describe("UpdateProject", () => {
     ) as HTMLInputElement;
     fireEvent.change(name, { target: { value: "Some name" } });
 
-    const updateButton = getByText("Update");
-    await act(async () => updateButton.click());
+    const number = container.querySelector(
+      "input[name=projectNumber]"
+    ) as HTMLInputElement;
+    fireEvent.change(number, { target: { value: "Some number" } });
+
+    const createButton = getByText("Create");
+    await act(async () => createButton.click());
     expect(fetchMock).toHaveBeenCalledWith(
-      "https://dev-api.bentley.com/projects/de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b",
+      "https://dev-api.bentley.com/projects",
       {
-        method: "PATCH",
+        method: "POST",
         headers: { Authorization: "dd", Prefer: "return=representation" },
         body: JSON.stringify({
           displayName: "Some name",
-          projectNumber: "Initial number",
+          projectNumber: "Some number",
         }),
       }
     );
