+++ packages/app/src/routers/CRUDRouter/components/delete-project/DeleteProject.test.tsx
@@ -1,6 +1,8 @@
 /*---------------------------------------------------------------------------------------------
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
+ *
+ * This code is for demonstration purposes and should not be considered production ready.
  *--------------------------------------------------------------------------------------------*/
 import { act, render } from "@testing-library/react";
 import React from "react";
