+++ packages/app/src/routers/CRUDRouter/components/delete-project/DeleteProject.tsx
@@ -1,6 +1,8 @@
 /*---------------------------------------------------------------------------------------------
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
+ *
+ * This code is for demonstration purposes and should not be considered production ready.
  *--------------------------------------------------------------------------------------------*/
 import SvgWarning from "@itwin/itwinui-icons-react/cjs/icons/StatusWarning";
 import {
