+++ packages/app/src/routers/CRUDRouter/components/update-project/UpdateProject.test.tsx
@@ -1,14 +1,16 @@
 /*---------------------------------------------------------------------------------------------
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
+ *
+ * This code is for demonstration purposes and should not be considered production ready.
  *--------------------------------------------------------------------------------------------*/
 import { toaster } from "@itwin/itwinui-react";
 import { act, fireEvent, render } from "@testing-library/react";
 import React from "react";
 
-import { CreateProject } from "./CreateProject";
+import { UpdateProject } from "./UpdateProject";
 
-describe("CreateProject", () => {
+describe("UpdateProject", () => {
   const mockedProject = { project: { id: "dd", displayName: "name" } };
   const fetchMock = jest.fn(() =>
     Promise.resolve({
@@ -26,44 +28,44 @@ describe("CreateProject", () => {
     jest.clearAllMocks();
   });
 
-  it("should create a project", async () => {
+  it("should update a project", async () => {
     const successMock = jest.fn();
     toaster.positive = jest.fn();
 
     const { getByText, container } = render(
-      <CreateProject
+      <UpdateProject
         accessToken="dd"
+        projectId="de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b"
         onSuccess={successMock}
         apiOverrides={{ serverEnvironmentPrefix: "dev" }}
+        initialProject={{
+          displayName: "Initial name",
+          projectNumber: "Initial number",
+        }}
       />
     );
 
     const name = container.querySelector(
       "input[name=displayName]"
     ) as HTMLInputElement;
-    fireEvent.change(name, { target: { value: "Some name" } });
+    fireEvent.change(name, { target: { value: "Some other name" } });
 
-    const number = container.querySelector(
-      "input[name=projectNumber]"
-    ) as HTMLInputElement;
-    fireEvent.change(number, { target: { value: "Some number" } });
-
-    const createButton = getByText("Create");
-    await act(async () => createButton.click());
+    const updateButton = getByText("Update");
+    await act(async () => updateButton.click());
     expect(fetchMock).toHaveBeenCalledWith(
-      "https://dev-api.bentley.com/projects",
+      "https://dev-api.bentley.com/projects/de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b",
       {
-        method: "POST",
+        method: "PATCH",
         headers: { Authorization: "dd", Prefer: "return=representation" },
         body: JSON.stringify({
-          displayName: "Some name",
-          projectNumber: "Some number",
+          displayName: "Some other name",
+          projectNumber: "Initial number",
         }),
       }
     );
     expect(successMock).toHaveBeenCalledWith(mockedProject);
     expect(toaster.positive).toHaveBeenCalledWith(
-      "Project created successfully.",
+      "Project updated successfully.",
       {
         hasCloseButton: true,
       }
@@ -77,10 +79,15 @@ describe("CreateProject", () => {
     toaster.negative = jest.fn();
 
     const { getByText, container } = render(
-      <CreateProject
+      <UpdateProject
         accessToken="dd"
+        projectId="de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b"
         onError={errorMock}
         apiOverrides={{ serverEnvironmentPrefix: "dev" }}
+        initialProject={{
+          displayName: "Initial name",
+          projectNumber: "Initial number",
+        }}
       />
     );
 
@@ -89,21 +96,16 @@ describe("CreateProject", () => {
     ) as HTMLInputElement;
     fireEvent.change(name, { target: { value: "Some name" } });
 
-    const number = container.querySelector(
-      "input[name=projectNumber]"
-    ) as HTMLInputElement;
-    fireEvent.change(number, { target: { value: "Some number" } });
-
-    const createButton = getByText("Create");
-    await act(async () => createButton.click());
+    const updateButton = getByText("Update");
+    await act(async () => updateButton.click());
     expect(fetchMock).toHaveBeenCalledWith(
-      "https://dev-api.bentley.com/projects",
+      "https://dev-api.bentley.com/projects/de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b",
       {
-        method: "POST",
+        method: "PATCH",
         headers: { Authorization: "dd", Prefer: "return=representation" },
         body: JSON.stringify({
           displayName: "Some name",
-          projectNumber: "Some number",
+          projectNumber: "Initial number",
         }),
       }
     );
@@ -111,7 +113,7 @@ describe("CreateProject", () => {
     expect(
       toaster.negative
     ).toHaveBeenCalledWith(
-      "Could not create a project. Please try again later.",
+      "Could not update a project. Please try again later.",
       { hasCloseButton: true }
     );
   });
@@ -123,10 +125,15 @@ describe("CreateProject", () => {
     toaster.negative = jest.fn();
 
     const { getByText, container } = render(
-      <CreateProject
+      <UpdateProject
         accessToken="dd"
+        projectId="de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b"
         onError={errorMock}
         apiOverrides={{ serverEnvironmentPrefix: "dev" }}
+        initialProject={{
+          displayName: "Initial name",
+          projectNumber: "Initial number",
+        }}
       />
     );
 
@@ -135,21 +142,16 @@ describe("CreateProject", () => {
     ) as HTMLInputElement;
     fireEvent.change(name, { target: { value: "Some name" } });
 
-    const number = container.querySelector(
-      "input[name=projectNumber]"
-    ) as HTMLInputElement;
-    fireEvent.change(number, { target: { value: "Some number" } });
-
-    const createButton = getByText("Create");
-    await act(async () => createButton.click());
+    const updateButton = getByText("Update");
+    await act(async () => updateButton.click());
     expect(fetchMock).toHaveBeenCalledWith(
-      "https://dev-api.bentley.com/projects",
+      "https://dev-api.bentley.com/projects/de47c5ad-5657-42b8-a2bc-f2b8bf84cd4b",
       {
-        method: "POST",
+        method: "PATCH",
         headers: { Authorization: "dd", Prefer: "return=representation" },
         body: JSON.stringify({
           displayName: "Some name",
-          projectNumber: "Some number",
+          projectNumber: "Initial number",
         }),
       }
     );
