+++ packages/app/src/routers/CRUDRouter/useCreateProjectAction.tsx
@@ -1,32 +1,32 @@
 /*---------------------------------------------------------------------------------------------
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
+ *
+ * This code is for demonstration purposes and should not be considered production ready.
  *--------------------------------------------------------------------------------------------*/
-import { ProjectFull } from "@itwin/imodel-browser-react";
-import { SvgEdit } from "@itwin/itwinui-icons-react";
+import { SvgAdd } from "@itwin/itwinui-icons-react";
+import { Button } from "@itwin/itwinui-react";
 import { NavigateFn } from "@reach/router";
 import React from "react";
 
-type EditProjectActionOptions = {
+export type CreateProjectActionOptions = {
   /**
    * Must be the "relative" navigate function, coming from route props.
    */
   navigate: NavigateFn | undefined;
 };
 
-export const useEditProjectAction = ({
+export const useCreateProjectAction = ({
   navigate,
-}: EditProjectActionOptions) => {
+}: CreateProjectActionOptions) => {
   return {
-    editAction: React.useMemo(
-      () => ({
-        key: "edit",
-        icon: <SvgEdit />,
-        onClick: (project: ProjectFull) =>
-          void navigate?.(`project/${project.id}/edit-project`),
-        children: "Edit project",
-      }),
-      [navigate]
+    createIconButton: (
+      <Button
+        onClick={() => void navigate?.(`create-project`)}
+        startIcon={<SvgAdd />}
+      >
+        New project
+      </Button>
     ),
   };
 };
