+++ packages/app/src/routers/SynchronizationRouter/components/BridgeIcon.scss
