+++ packages/app/src/routers/SynchronizationRouter/components/ConnectionsTable.scss
