+++ packages/app/src/routers/SynchronizationRouter/components/DebugTools.scss
