+++ packages/app/src/routers/SynchronizationRouter/components/DetailedStatus.scss
