+++ packages/app/src/routers/SynchronizationRouter/components/SkeletonCell.scss
