+++ packages/app/src/routers/SynchronizationRouter/components/UploadPanel.scss
